+++ lib/src/analyzers/lint_analyzer/rules/rules_list/format_comment/models/comment_info.dart
@@ -1,0 +1,8 @@
+part of '../format_comment_rule.dart';
+
+class _CommentInfo {
+  final Token token;
+  final _CommentType type;
+
+  _CommentInfo(this.type, this.token);
+}