+++ lib/src/analyzers/lint_analyzer/rules/rules_list/format_comment/models/comment_type.dart
@@ -1,0 +1,3 @@
+part of '../format_comment_rule.dart';
+
+enum _CommentType { base, documentation }