+++ test/src/analyzers/lint_analyzer/rules/rules_list/format_comment/examples/example.dart
@@ -1,0 +1,18 @@
+// With start space without dot
+class Test {
+  //Without start space without dot
+  Test() {
+    // with start space with dot.
+  }
+
+  /// With start space without dot
+  function() {
+    /// with start space with dot.
+  }
+// ignore:
+}
+
+// ignore_for_file:
+var a;
+
+/// [WidgetModel] for widget.